requirements.txt
aiohttp==3.8.1
aioresponses==0.7.3
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coverage==6.3.1
Django==4.0.2
djangorestframework==3.13.1
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
multidict==6.0.2
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.0
pytest-asyncio==0.18.1
pytest-django==4.5.2
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarl==1.7.2
